% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celsius2fahrenheit.R
\name{c.2.f}
\alias{c.2.f}
\title{A Function to Provide Environmental Context to Data}
\usage{
\method{c}{2.f}(c)
}
\arguments{
\item{c:}{class: numeric; description: value in degrees Celsius}
}
\description{
A Function to Provide Environmental Context to Data
}
