% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory_buffer.R
\name{clear_memory}
\alias{clear_memory}
\title{Clear all messages}
\usage{
clear_memory(buffer)
}
\arguments{
\item{buffer}{A MemoryBuffer.}
}
\value{
The same buffer (invisible), emptied.
}
\description{
Clear all messages
}
