% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context-audit.R
\name{context_audit}
\alias{context_audit}
\title{Audit a contextual object}
\usage{
context_audit(x, ...)
}
\arguments{
\item{x}{A \code{contextual} object.}

\item{...}{Unused.}
}
\value{
A list containing traceability and check metadata.
}
\description{
Audit a contextual object
}
