% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_airnow_api_call.R
\name{create_airnow_api_call}
\alias{create_airnow_api_call}
\title{contextR}
\usage{
create_airnow_api_call(API_KEY, lat, lng, ts, distance = 10)
}
\arguments{
\item{API_KEY}{class: string}

\item{lat}{class: numeric}

\item{lng}{class: numeric}

\item{ts}{class: string (ISO format: 2019-10-23T16:00:00)}

\item{distance}{class: numeric (miles)}
}
\description{
contextR
}
\examples{
create_airnow_api_call(API_KEY, lat, lng, ts, distance=10)
}
