% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fcc_api_call.R
\name{create_fcc_api_call}
\alias{create_fcc_api_call}
\title{contextR}
\usage{
create_fcc_api_call(lat, lng)
}
\arguments{
\item{lat}{class: numeric}

\item{lng}{class: numeric}
}
\description{
contextR
}
\examples{
create_fcc_api_call(lat, lng)
}
