% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistence.R
\name{disable_autosave}
\alias{disable_autosave}
\title{Disable autosave for a MemoryBuffer}
\usage{
disable_autosave(buffer)
}
\arguments{
\item{buffer}{MemoryBuffer}
}
\description{
Disable autosave for a MemoryBuffer
}
