% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_airquality_context.R
\name{get_airquality_context}
\alias{get_airquality_context}
\title{contextR}
\usage{
get_airquality_context(
  data,
  rate_limit = 8.6,
  save_results = T,
  results_dir = ""
)
}
\arguments{
\item{data}{class: data.frame}

\item{rate_limit}{class: numeric}

\item{save_results}{class: boolean}

\item{results_dir}{class: character}
}
\description{
contextR
}
\examples{
get_weather_context(api_call, hash="md5")
}
