% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lunar_context_from_date.R
\name{get_lunar_context_from_date}
\alias{get_lunar_context_from_date}
\title{contextR}
\usage{
get_lunar_context_from_date(date, ...)
}
\arguments{
\item{date}{class: POSIXct}
}
\description{
contextR
}
\examples{
get_lunar_context_from_date(date)
}
