% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory_buffer.R
\name{get_memory}
\alias{get_memory}
\title{Get messages as a tibble}
\usage{
get_memory(buffer)
}
\arguments{
\item{buffer}{A MemoryBuffer.}
}
\value{
tibble with columns: role, content, timestamp, name, meta.
}
\description{
Get messages as a tibble
}
