% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory_buffer.R
\name{get_save_path}
\alias{get_save_path}
\title{Get the current autosave target path}
\usage{
get_save_path(buffer)
}
\arguments{
\item{buffer}{A \code{MemoryBuffer}.}
}
\description{
Get the current autosave target path
}
