% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_season_from_date.R
\name{get_season_from_date}
\alias{get_season_from_date}
\title{contextR}
\usage{
get_season_from_date(date, hemi = "northern", use_pkg = T)
}
\arguments{
\item{date}{class: POSIXct}

\item{hemi}{class: string}

\item{use_pkg}{class: boolean}
}
\description{
contextR
}
\examples{
get_season_from_date(date, hemi="northern", use_pkg = T)
}
