% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_context.R
\name{get_weather_context}
\alias{get_weather_context}
\title{contextR}
\usage{
get_weather_context(api_call, hash = "md5")
}
\arguments{
\item{api_call}{class: string}

\item{hash}{class: string}
}
\description{
contextR
}
\examples{
get_weather_context(api_call, hash="md5")
}
