% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_zodiac_sign_from_dob.R
\name{get_zodiac_sign_from_date}
\alias{get_zodiac_sign_from_date}
\title{contextR}
\usage{
get_zodiac_sign_from_date(date)
}
\arguments{
\item{date}{class: POSIXct}
}
\description{
contextR
}
\examples{
get_zodiac_sign_from_date(date)
}
