% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEnvironmentalData.R
\name{getEnviromentalData}
\alias{getEnviromentalData}
\title{A Function to Provide Environmental Context to Data}
\usage{
getEnviromentalData(df, nLunar = 8, returnAll = F)
}
\arguments{
\item{df}{class: data.frame description: must have columns labelled identically to: gps_lat, gps_long, user_id, date}

\item{nLunar}{class: integer; description: number of lunar phase factors to return}

\item{returnAll}{class: boolean; T = returns list with many steps along the way to creating final dataframe; F = returns dataframe}
}
\description{
A Function to Provide Environmental Context to Data
}
\examples{
getEnviromentalData(df,nLunar=8,returnAll=F) # just give me my dataset back with contextual data
getEnviromentalData(df,nLunar=8,returnAll=T) # I want to see all the behind the scenes data and do more
}
