% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_api_call.R
\name{hash_api_call}
\alias{hash_api_call}
\title{contextR}
\usage{
hash_api_call(api_call, algo = "md5")
}
\arguments{
\item{api_call}{class: string}

\item{hash}{class: string}
}
\description{
contextR
}
\examples{
hash_api_call(api_call, hash="md5")
}
