% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory_buffer.R
\name{last_assistant_text}
\alias{last_assistant_text}
\title{Get the last assistant message from the buffer as plain text}
\usage{
last_assistant_text(buffer)
}
\arguments{
\item{buffer}{A \code{MemoryBuffer}.}
}
\description{
Get the last assistant message from the buffer as plain text
}
