% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lunar2DF.R
\name{lunar2df}
\alias{lunar2df}
\title{A Function to Provide Environmental Context to Data}
\usage{
lunar2df(df, id = "user_id", date_c = "date", date_f = "\%Y/\%m/\%d",
  lat = "gps_lat", long = "gps_long", nLunar = 8)
}
\arguments{
\item{df}{class: data.frame description: must have columns labelled identically to: gps_lat, gps_long, user_id, date}

\item{date_c}{class: string; variable name for column with date data}

\item{date_f}{class: string; date format}

\item{lat}{class: string; variable name for column with GPS Latitude data}

\item{long}{class: string; variable name for column with GPS Longitude data}
}
\description{
A Function to Provide Environmental Context to Data
}
\examples{
lunar2df(df,date_c="date", date_f="\%Y/\%m/\%d", id='user_id', lat='gps_lat', long='gps_long')
}
