% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory_buffer.R
\name{memory_persistence}
\alias{memory_persistence}
\alias{save_memory}
\alias{load_memory}
\title{Save and load memory buffers}
\usage{
save_memory(buffer, file)

load_memory(file)
}
\arguments{
\item{buffer}{A MemoryBuffer to save.}

\item{file}{Path to an .rds file.}
}
\description{
Save and load memory buffers
}
