% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory_buffer.R
\name{memory_prompt}
\alias{memory_prompt}
\title{Compose a final prompt with current memory + a new user prompt}
\usage{
memory_prompt(buffer, user_prompt, format = c("plain", "ellmer"))
}
\arguments{
\item{buffer}{A MemoryBuffer.}

\item{user_prompt}{The new user prompt/question.}

\item{format}{One of "plain","ellmer".}
}
\value{
Combined prompt string.
}
\description{
Compose a final prompt with current memory + a new user prompt
}
