% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context-cor-knn-inputs.R
\name{new_context_knn_input}
\alias{new_context_knn_input}
\title{Create KNN input for \code{context()}}
\usage{
new_context_knn_input(data, class_var, x1, x2, k = 5)
}
\arguments{
\item{data}{Data frame.}

\item{class_var}{Class column name.}

\item{x1}{First predictor name.}

\item{x2}{Second predictor name.}

\item{k}{Number of neighbors.}
}
\value{
An object of class \code{context_knn_input}.
}
\description{
Create KNN input for \code{context()}
}
