% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory_buffer.R
\name{set_autosave}
\alias{set_autosave}
\title{Enable/disable autosave and/or update path}
\usage{
set_autosave(buffer, enabled = TRUE, save_dir = NULL, save_file = NULL)
}
\arguments{
\item{buffer}{MemoryBuffer}

\item{enabled}{logical}

\item{save_dir}{optional new dir}

\item{save_file}{optional new file name}
}
\description{
Enable/disable autosave and/or update path
}
