% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u_gps_distance_origin.R
\name{u_gps_distance_origin}
\alias{u_gps_distance_origin}
\title{contextR}
\usage{
u_gps_distance_origin(homepoint, lat2, lng2, return_type = "meters")
}
\arguments{
\item{homepoint}{class: vector(lat, lng)}

\item{lat2}{class: numeric}

\item{lng2}{class: numeric}

\item{return_type}{class: string}
}
\description{
contextR
}
\examples{
u_gps_distance_origin(homepoint, lat2, lon2, return_type = "meters")
}
