% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u_gps_pair_distance.R
\name{u_gps_pair_distance}
\alias{u_gps_pair_distance}
\title{contextR}
\usage{
u_gps_pair_distance(lat1, lon1, lat2, lon2, return_type = "meters")
}
\arguments{
\item{lat1}{class: numeric}

\item{lat2}{class: numeric}

\item{return_type}{class: string}

\item{lng1}{class: numeric}

\item{lng2}{class: numeric}
}
\description{
contextR
}
\examples{
u_gps_pair_distance(lat1, lng1, lat2, lng2, return_type = "meters")
}
