% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u_meters_to_miles.R
\name{u_meters_to_miles}
\alias{u_meters_to_miles}
\title{contextR}
\usage{
u_meters_to_miles(meters)
}
\arguments{
\item{meters}{class: numeric}
}
\description{
contextR
}
\examples{
u_meters_to_miles(meters)
}
