% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u_zipcode_to_geo.R
\name{u_zipcode_to_geo}
\alias{u_zipcode_to_geo}
\title{contextR}
\usage{
u_zipcode_to_geo(zip, ...)
}
\arguments{
\item{zip}{class: numeric}
}
\description{
contextR
}
\examples{
u_zipcode_to_geo(zip)
}
