% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather2DF.R
\name{weather2df}
\alias{weather2df}
\title{A Function to Provide Environmental Context to Data}
\usage{
weather2df(df, returnAll = F)
}
\arguments{
\item{df}{class: data.frame description: must have columns labelled identically to: gps_lat, gps_long, user_id, date}

\item{returnAll}{class: boolean; T = returns list with many steps along the way to creating final dataframe; F = returns dataframe}

\item{nLunar}{class: integer; description: number of lunar phase factors to return}
}
\description{
A Function to Provide Environmental Context to Data
}
\examples{
weather2df(df,returnAll=F) # just give me my dataset back with contextual data
weather2df(df,returnAll=T) # I want to see all the behind the scenes data and do more
}
